+++ apps/cowswap-frontend/src/modules/twap/hooks/useTwapOrderCreationContext.ts
@@ -1,8 +1,8 @@
 import { useMemo } from 'react'
 
-import { ComposableCoW, Erc20 } from '@cowprotocol/abis'
 import { useTradeSpenderAddress } from '@cowprotocol/balances-and-allowances'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
+import { ComposableCoW, Erc20 } from '@cowprotocol/cowswap-abis'
 import { CurrencyAmount, Token } from '@uniswap/sdk-core'
 
 import { Nullish } from 'types'
@@ -13,7 +13,6 @@ import { useNeedsZeroApproval } from 'modules/zeroApproval'
 
 import { useTokenContract } from 'common/hooks/useContract'
 import { useNeedsApproval } from 'common/hooks/useNeedsApproval'
-
 
 export interface TwapOrderCreationContext {
   composableCowContract: ComposableCoW