+++ libs/wallet/src/web3-react/updater.ts
@@ -68,20 +68,13 @@ function useWalletDetails(account?: string, standaloneMode?: boolean): WalletDet
   }, [isSmartContractWallet, isSafeApp, walletName, icon, ensName])
 }
 
-// TODO: Break down this large function into smaller functions
-// eslint-disable-next-line max-lines-per-function
 function useSafeInfo(walletInfo: WalletInfo): GnosisSafeInfo | undefined {
-  const { provider } = useWeb3React()
   const { account, chainId } = walletInfo
   const [safeInfo, setSafeInfo] = useState<GnosisSafeInfo>()
   const safeAppsSdk = useSafeAppsSdk()
 
-  // TODO: Break down this large function into smaller functions
-  // eslint-disable-next-line max-lines-per-function
   useEffect(() => {
-    // TODO: Add proper return type annotation
-    // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-    const updateSafeInfo = async () => {
+    const updateSafeInfo: () => Promise<void> = async () => {
       if (safeAppsSdk) {
         try {
           const appsSdkSafeInfo = await safeAppsSdk.safe.getInfo()
@@ -93,7 +86,7 @@ function useSafeInfo(walletInfo: WalletInfo): GnosisSafeInfo | undefined {
               chainId,
               threshold,
               owners,
-              nonce,
+              nonce: Number(nonce),
               isReadOnly,
             }
           })
@@ -102,9 +95,9 @@ function useSafeInfo(walletInfo: WalletInfo): GnosisSafeInfo | undefined {
           setSafeInfo(undefined)
         }
       } else {
-        if (chainId && account && provider) {
+        if (chainId && account) {
           try {
-            const _safeInfo = await getSafeInfo(chainId, account, provider)
+            const _safeInfo = await getSafeInfo(chainId, account)
             const { address, threshold, owners, nonce } = _safeInfo
             setSafeInfo((prevSafeInfo) => ({
               ...prevSafeInfo,
@@ -112,7 +105,8 @@ function useSafeInfo(walletInfo: WalletInfo): GnosisSafeInfo | undefined {
               address,
               threshold,
               owners,
-              nonce,
+              // Time to time Safe sends a string or a number
+              nonce: Number(nonce),
               isReadOnly: false,
             }))
           } catch {
@@ -147,7 +141,7 @@ function useSafeInfo(walletInfo: WalletInfo): GnosisSafeInfo | undefined {
       clearInterval(longSafeInfoInterval !== null ? longSafeInfoInterval : undefined)
       longSafeInfoInterval = null
     }
-  }, [setSafeInfo, chainId, account, provider, safeAppsSdk])
+  }, [chainId, account, safeAppsSdk])
 
   return safeInfo
 }